fix markup type

+++ src/Entities/Catalog/SkuPrice.php
@@ -8,7 +8,7 @@ use Orkestra\Entities\AbstractEntity;
  * @property-read string $itemId
  * @property-read int|null $listPrice
  * @property-read int $costPrice
- * @property-read int $markup
+ * @property-read float $markup
  * @property-read int $basePrice
  * @property-read array<int,array{tradePolicyId:string,value:float,listPrice:float|null,minQuantity:int,dateRange:array{from:string,to:string}}> $fixedPrices
  */
@@ -21,7 +21,7 @@ class SkuPrice extends AbstractEntity
 		protected string $itemId,
 		protected ?int $listPrice,
 		protected int $costPrice,
-		protected int $markup,
+		protected float $markup,
 		protected int $basePrice,
 		protected array $fixedPrices,	
 	) {
